expanded cast parameters to blindly try and avoid numeric overflows

+++ filesets/dims/_dim_user.dbc.coa.sql
@@ -89,11 +89,11 @@ create volatile table dim_user as
                  from (
                         select UserName as Active_UserName
                         ,zeroifnull(sum(cast(Statements as BigInt))) as Query_Cnt
-                        ,zeroifnull(avg(cast(NumSteps as Int))) as Query_Complexity_Score
-                        ,zeroifnull(sum(cast(ParserCPUTime+AMPCPUTime as Decimal(18,2)))) as CPU_Sec
-                        ,zeroifnull(cast(sum(ReqIOKB(BigInt))/1e6 as Decimal(18,3))) as IOGB
-                        ,zeroifnull(sum(cast(TotalFirstRespTime as Decimal(18,6)))) as Runtime_Sec
-                        ,zeroifnull(sum(cast((case when ErrorCode not in(0,3158) then 1 else 0 end) as decimal(9,0)))) as Error_Cnt
+                        ,zeroifnull(avg(cast(NumSteps as BigInt))) as Query_Complexity_Score
+                        ,zeroifnull(sum(cast(ParserCPUTime+AMPCPUTime as Decimal(32,2)))) as CPU_Sec
+                        ,zeroifnull(cast(sum(ReqIOKB(BigInt))/1e6 as Decimal(32,3))) as IOGB
+                        ,zeroifnull(sum(cast(TotalFirstRespTime as Decimal(32,6)))) as Runtime_Sec
+                        ,zeroifnull(sum(cast((case when ErrorCode not in(0,3158) then 1 else 0 end) as decimal(32,0)))) as Error_Cnt
                         from dbc.QryLogV
                         where cast(starttime as DATE) between {startdate} and {enddate}
                         group by 1
@@ -101,8 +101,8 @@ create volatile table dim_user as
                         select UserName as Active_UserName
                         ,zeroifnull(sum(cast(QueryCount as BigInt))) as Query_Cnt
                         ,null as Query_Complexity_Score
-                        ,zeroifnull(sum(cast(ParserCPUTime+AMPCPUTime as Decimal(18,2)))) as CPU_Sec
-                        ,zeroifnull(cast(sum(ReqPhysIOKB(BigInt))/1e6 as Decimal(18,3))) as IOGB
+                        ,zeroifnull(sum(cast(ParserCPUTime+AMPCPUTime as Decimal(32,2)))) as CPU_Sec
+                        ,zeroifnull(cast(sum(ReqPhysIOKB(BigInt))/1e6 as Decimal(32,3))) as IOGB
                         ,0 as Runtime_Sec
                         ,0 as Error_Cnt
                         from dbc.QryLogSummaryV
